move WaveWatch3Downloader into downloader module
