+++ wavewatch3/__init__.py
@@ -1,6 +1,6 @@
 from ._version import __version__
 from .bmi import BmiWaveWatch3
-from .wavewatch3 import WaveWatch3Downloader
+from .downloader import WaveWatch3Downloader
 
 
 __all__ = ["__version__", "BmiWaveWatch3", "WaveWatch3Downloader"]
